+++ src/gracy/_replay/_storages/_pymongo_storage.py
@@ -8,7 +8,7 @@ import httpx
 
 from gracy.exceptions import GracyReplayRequestNotFound
 
-from .._storages import GracyReplayStorage
+from ._base import GracyReplayStorage
 
 try:
     import pymongo
